app launch issue in local: fixed + refacto context provide in pages: done
