+++ .nvmrc
@@ -1,1 +1,2 @@
+20.19.5
 20
