+++ gatsby-browser.js
@@ -1,0 +1,6 @@
+import React from "react"
+import Layout from "./src/components/layout"
+
+export const wrapPageElement = ({ element, props }) => {
+  return <Layout {...props}>{element}</Layout>
+}
