+++ package.json
@@ -4,6 +4,10 @@
   "description": "An optimized communication tool for hotel personal staff.",
   "version": "0.1.0",
   "author": "David Simba <david.simba@mysweethotel.com>",
+  "engines": {
+    "node": ">=20 <21"
+  },
+  "packageManager": "npm@10.8.2",
   "dependencies": {
     "@date-io/date-fns": "^1.3.13",
     "@date-io/moment": "^1.3.13",
@@ -38,7 +42,7 @@
     "gatsby-firesource": "^2.0.3",
     "gatsby-plugin-google-fonts": "^1.0.1",
     "gatsby-plugin-image": "^3.13.1",
-    "gatsby-plugin-manifest": "^4.7.0",
+    "gatsby-plugin-manifest": "^5.13.1",
     "gatsby-plugin-offline": "^6.13.1",
     "gatsby-plugin-react-helmet": "^6.13.1",
     "gatsby-plugin-react-i18next": "^3.0.1",
@@ -112,7 +116,9 @@
   "license": "MIT",
   "scripts": {
     "build": "gatsby build",
+    "clean:gatsby": "gatsby clean",
     "develop": "gatsby develop",
+    "develop:clean": "npm run clean:gatsby && gatsby develop",
     "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
     "start": "npm run develop",
     "serve": "gatsby serve",
@@ -125,5 +131,11 @@
   },
   "bugs": {
     "url": "https://github.com/gatsbyjs/gatsby/issues"
+  },
+  "overrides": {
+    "lmdb": "^3.2.7"
+  },
+  "optionalDependencies": {
+    "@parcel/watcher-darwin-arm64": "^2.5.1"
   }
 }
