+++ src/Firebase/context.js
@@ -1,5 +1,7 @@
 import React from "react";
 
-const FirebaseContext = React.createContext(null);
+// Default value is a non-null object so consumers can safely destructure
+// even if a Provider is missing/misconfigured.
+const FirebaseContext = React.createContext({ user: null, userDB: null });
 
 export default FirebaseContext;
