+++ src/pages/404.js
@@ -1,5 +1,4 @@
 import React from "react"
-import Layout from '../components/layout'
 import SEO from "../components/seo"
 
 const NotFoundPage = () => (
