+++ src/pages/chat.js
@@ -1,6 +1,5 @@
-import React, { useEffect, useState } from "react"
-import Loader from '../helper/common/mshLoader'
-import {FirebaseContext, db, auth} from '../Firebase'
+import React, { useContext } from "react"
+import {FirebaseContext} from '../Firebase'
 import Chat from '../components/section/communIzi'
 import PhoneChat from '../components/section/phoneCommunIzi'
 import Navigation from '../components/section/navigation'
@@ -9,31 +8,14 @@ import { useTranslation } from "react-i18next"
 import ToolBar from "../components/section/toolbar"
 
 const ChatPage = () => {
-  const [hide, setHide] = useState("flex")
-  const [userDB, setUserDB] = useState(null)
-  const [user, setUser] = useState(null)
+  const { userDB, setUserDB, user, setUser } = useContext(FirebaseContext)
   const { t } = useTranslation()
-
-  useEffect(() => {     
-    if(!user && !userDB){
-      const userStorage = JSON.parse(sessionStorage.getItem('userStorage'))
-    const userAuth = JSON.parse(sessionStorage.getItem('userAuth'))
-    
-    setUser(userAuth)
-    setUserDB(userStorage)
-    return setHide("none")
-    }
-        
-}, [user, userDB])
 
 const isBrowser = () => typeof window !== "undefined"
 
   return(
-    <FirebaseContext.Provider value={{ userDB, setUserDB, user, setUser }}> 
+    <> 
         <div className="landscape-display"></div>
-        <div style={{position: "absolute", zIndex: "9", width: "100%"}}> 
-          <Loader hide={hide} />
-        </div>     
         {!!user && !!userDB &&
         <Navigation user={user} userDB={userDB} />}  
           <div style={{
@@ -54,7 +36,7 @@ const isBrowser = () => typeof window !== "undefined"
               <PhoneChat userDB={userDB} user={user} />}
             </div>}
           </div>
-    </FirebaseContext.Provider>
+    </>
   )
 }
