+++ src/pages/crm.js
@@ -1,35 +1,16 @@
-import React, { useEffect, useState } from "react"
-import Loader from '../helper/common/mshLoader'
-import {FirebaseContext, db, auth} from '../Firebase'
+import React, { useContext } from "react"
+import {FirebaseContext} from '../Firebase'
 import GuestDatabase from '../components/section/guestDatabase'
 import Navigation from '../components/section/navigation'
 import { withTrans } from '../../i18n/withTrans'
 import ToolBar from "../components/section/toolbar"
-import { parse } from "url"
 
 const UserDatabase = () => {
-  const [hide, setHide] = useState("flex")
-  const [userDB, setUserDB] = useState(null)
-  const [user, setUser] = useState(null)
-
-  useEffect(() => {     
-    if(!user && !userDB){
-      const userStorage = JSON.parse(sessionStorage.getItem('userStorage'))
-    const userAuth = JSON.parse(sessionStorage.getItem('userAuth'))
-    
-    setUser(userAuth)
-    setUserDB(userStorage)
-    return setHide("none")
-    }
-        
-}, [user, userDB])
+  const { userDB, setUserDB, user, setUser } = useContext(FirebaseContext)
 
   return(
-    <FirebaseContext.Provider value={{ userDB, setUserDB, user, setUser }}> 
+    <> 
         <div className="landscape-display"></div>
-        <div style={{position: "absolute", zIndex: "9", width: "100%"}}> 
-          <Loader hide={hide} />
-        </div>   
         {!!user && !!userDB &&
         <Navigation user={user} userDB={userDB} />}    
         <div style={{
@@ -39,7 +20,7 @@ const UserDatabase = () => {
           {!!user && !!userDB &&
           <GuestDatabase user={user} userDB={userDB} />}
         </div>
-    </FirebaseContext.Provider>
+    </>
   )
 }
